requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ffmpeg-python==0.2.0
Flask==3.0.3
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydub==0.25.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
